Organized source files in folder structure
